Fixed music max-volume upon track start & did basic Twitch work
Before when a song started, the volume would be multiplied by 1000, which would cause it to be very loud, so I fixed that.

Also now every time the bot starts, a new Twitch access token is generated and automatically set, meaning it doesn't have to be changed every period of time.